+++ src/DropdownTest.jsx
@@ -8,6 +8,8 @@
  * =========================================================
  * 2025/12/16               1.0.0     ITA     Genesis.
  * 2026/01/11  2026/01/16   1.0.1     ITA     Use of the Dropdown component no longer requires a Collections context provider.
+ * 2026/01/20  2026/01/27   1.0.2     ITA     Added a test for selReset prop.
+ * =========================================================
 */
 
 /** VERY IMPORTANT!!!
@@ -107,6 +109,7 @@ export default function DropdownTest() {
                           label={'Topics'}
                           data={topics}
                           selected={selectedTopic}
+                          selReset={false}
                           onItemSelected={topicSelected}
                           dropdownStyle={{color: '#000', backgroundColor: '#66ff66'}} />               
             </div>